[Chore] 할부 정보 커스텀 훅 분리
